Added sample .gitignore, build and code generation files
./build-office-files.sh
This script is supposed to be run after updating the source code in the repository, for example in a post-checkout git hook. The script builds the office files from source (where source is defined as the uncompressed office files).

./generate-office-source.sh
This script is supposed to be run after making changes to the office files and before commiting the changes to git. It is possible to run as a pre-commit hook.

.gitignore
Added this to show that the binaries are supposed to be ignored in the repository. The reason is that the uncompressed data should be seen as the source of truth and thus should be the only code checked in; the binary office files are instead seen as build results.