Pretty format xml files
Turns out that source controlling files with just single, huge lines is not going to help very much.

This commit introduces pretty formatting of the XML source, which should make source control easier. It relies on the `xmllint` tool which came preinstalled at least with git on Windows.

Also turns out that the office files build from a prettified source works well; no need to minify the files for the build at the moment.